liquid_backend/requirements_server.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
liquidtrading-python==0.1.3
pydantic>=2.6.0
python-multipart==0.0.6
